Refactor: Move GroupHom and related functions to Hom.ts
